refactor: refactor files structure
