+++ src/index.ts
@@ -1,3 +1,3 @@
-import './textbook/textbook.ts';
-import './textbook/navigation.ts';
-import './textbook/localStorageUtil.ts';
+import './components/textbook/textbook.ts';
+import './components/textbook/navigation.ts';
+import './components/textbook/localStorageUtil.ts';
