+++ src/sass/textbook.scss
@@ -134,7 +134,7 @@ ul {
   width: 40px;
   height: 40px;
   border: 2px solid #ffffff;
-  background-image: url("../textbook/sound.svg");
+  background-image: url("../assets/svg/sound.svg");
   border-radius: 100%;
   background-repeat: no-repeat;
   object-fit: cover;
